Added logging out and deleting account. Also changed email verification when changing email

+++ app/src/main/java/com/example/project/UserProfileActivity.java
@@ -167,6 +167,7 @@ public class UserProfileActivity extends AppCompatActivity {
                 warning.setPositiveButton("Logout", new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
+                        FirebaseAuth.getInstance().signOut();
                         Toast.makeText(getApplicationContext(), "Logging out", Toast.LENGTH_SHORT).show();
                         Intent i = new Intent();
                         i.setClass(getApplicationContext(), MainActivity.class);
@@ -193,11 +194,20 @@ public class UserProfileActivity extends AppCompatActivity {
                 warning.setPositiveButton("Delete", new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
-                        Toast.makeText(getApplicationContext(), "Your profile has been deleted", Toast.LENGTH_SHORT).show();
-                        currentuser.delete();
-                        Intent i = new Intent();
-                        i.setClass(getApplicationContext(), MainActivity.class);
-                        startActivity(i);
+                        if (user != null) {
+                            user.delete()
+                                    .addOnCompleteListener(task -> {
+                                        if (task.isSuccessful()) {
+                                            Toast.makeText(getApplicationContext(), "Your profile has been deleted", Toast.LENGTH_SHORT).show();
+                                            currentuser.delete();
+                                            Intent i = new Intent();
+                                            i.setClass(getApplicationContext(), MainActivity.class);
+                                            startActivity(i);
+                                        } else {
+                                            Log.e("DeleteUser", "Auth deletion failed: " + task.getException().getMessage());
+                                        }
+                                    });
+                        }
                     }
                 });
                 warning.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
